% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{shikyo_html}
\alias{shikyo_html}
\title{平均価格html取得}
\usage{
shikyo_html(formdata)
}
\arguments{
\item{formdata}{取得したいテーブルに対応した文字列を指定。
- "VEGI": 野菜
- "FRUIT": 果実
- "FLOWER": 花き}
}
\description{
平均価格html取得
}
